Split rbm_spin_phase into header and source files
